New fully scaling positioning (you can provide a minimum and maximum padding between events) Better seperation of components Added dependency to react-dimensions to eliminate own wrapper component
