+++ package.json
@@ -92,6 +92,7 @@
     "color": "^0.11.3",
     "radium": "^0.18.0",
     "react": "^15.0.2",
+    "react-dimensions": "^1.3.0",
     "react-icons": "^2.0.1",
     "react-motion": "^0.4.2"
   },
