+++ src/helpers.js
@@ -26,19 +26,21 @@ export const zip = rows => rows[0].map((_, c) => rows.map(row => row[c]));
 
 
 /**
- * Determines the minimum distance between events
- * @param {array} dates the array containing all the dates
- * @param {number} seperation the minimum seperation required.
- * @return {number} the minimum distance between events
+ * Determines the minimum and maximum distance between a list of dates
+ * @param {array} dates The array containing all the dates
+ * @return {{min: number, max: number}} The minimum and maximum distances
  */
-export const minDistanceEvents = (dates, seperation) => {
+export const dateDistanceExtremes = (dates) => {
   // determine the minimum distance among events
   const datePairs = zip([ dates.slice(0, -1), dates.slice(1) ]);
   const dateDistances = datePairs.map(([ x, y ]) => daydiff(x, y));
 
   // return the minimum distance between two dates but considering that all dates
   // are the same then return the provided minimum seperation.
-  return Math.max(Math.min.apply(null, dateDistances), seperation);
+  return {
+    min: Math.min.apply(null, dateDistances),
+    max: Math.max.apply(null, dateDistances),
+  }
 };
 
 
@@ -47,30 +49,30 @@ export const minDistanceEvents = (dates, seperation) => {
  * that set of dates.
  *
  * @param {dates} the array containing dates the dates
- * @param {maxSeperation} upper bound on the seperation of dates irrespective of how far they are.
- * @param {minDistance} lower bound on the seperation of dates irrespective of how close they are.
+ * @param {number} labelWidth The width the label is going to use
+ * @param {number} minEventPadding The minimum padding between events.
+ * @param {number} maxEventPadding The maximum padding between events.
+ * @param {number} startPadding The padding at the beginning of the timeline
  * @return {array} positioning information for dates from a given origin point
  */
 // the interface for this function is pure
-export const cummulativeSeperation = (dates, minDistance, minSeperation, maxSeperation) => {
+export const cummulativeSeperation = (dates, labelWidth, minEventPadding, maxEventPadding, startPadding) => {
   // Calculating the minimum seperation between events
-  const eventsMinLapse = minDistanceEvents(dates, minSeperation);
+  const dateExtremes = dateDistanceExtremes(dates);
+  const datesDiff = dateExtremes.max - dateExtremes.min;
+  const paddingDiff = maxEventPadding - minEventPadding;
+  const halfLabel = labelWidth / 2;
 
   // using dynamic programming to set up the distance from the origin of the timeline.
   const distances = new Array(dates.length);
-  distances[0] = minDistance;
+  distances[0] = startPadding;
 
   for (let index = 1; index < distances.length; index += 1) {
     const distance = daydiff(dates[index - 1], dates[index]);
-    // NOTE: for now just setting a hard limit on the value of normalised distanceNorm
-    // i.e. distances will grow linearly and reach a max point then stop to increase
-    // an elegent mathametical calculation opertunity here.
-    const distanceFromPrevious = Math.min(
-      Math.round(distance / eventsMinLapse) + 1, maxSeperation
-    );
+    // relative spacing according to min and max seperation
+    const seperation = Math.round((((distance - dateExtremes.min) * paddingDiff) / datesDiff) + minEventPadding);
     // the distance_from_origin(n) = distance_from_origin(n-1) + distance between n and n - 1.
-    distances[index] = distances[index - 1] + distanceFromPrevious * minDistance;
+    distances[index] = distances[index - 1] + labelWidth + seperation;
   }
   return distances;
 };
-
